chore: resolve type error
