+++ __mocks__/bounty.ts
@@ -1,7 +1,7 @@
-import { Submission } from "@/types/bounty";
-import { community } from "./community";
-import { challenge } from "./course";
-import { reward } from "./reward";
+// import { community, mockUser } from "./community";
+// import { challenge } from "./course";
+// import { Referral, UserReferral } from "@/types/community";
+// import { mockSubmission } from "./submission";
 
 export const mockTrailer = {
   summary: "trailer summary",
@@ -14,54 +14,6 @@ export const mockTrailer = {
   },
 };
 
-export const mockSubmission: Submission = {
-  length: 0,
-  id: "submission_id",
-  ref: "reference",
-  created_at: new Date("2022-05-01T12:00:00Z"),
-  updated_at: new Date("2022-05-01T12:00:00Z"),
-  link: "/submissions/reference",
-  community: community,
-  user_id: "user_id",
-  challenge: challenge,
-  text: "Submission",
-  reviewDeadline: new Date(),
-  metadata: {
-    evaluation: "",
-    applicableReward: {
-      ref: "",
-      amount: 0,
-      updated_at: "",
-      challenge: "",
-      created_at: "",
-      id: "",
-      type: "",
-      community: "",
-      timestamp: 0,
-      token: "",
-    },
-    reviewed: false,
-    feedbacks: 0,
-    language: "",
-  },
-  timestamp: 0,
-  user: {
-    id: "user_id",
-    ref: "ref",
-    created_at: new Date("2022-05-01T12:00:00Z"),
-    firstName: "John Doe",
-    displayName: "John Doe",
-    uid: "uuid-uuido-232-dex0232-2331",
-    joined: "14 days ago",
-    disabled: false,
-    reputation: 0,
-    username: "",
-    lastName: "",
-    emailVerified: false,
-    email: "",
-    kycStatus: "",
-  },
-  reviewable: false,
-  status: "",
-  reward: reward,
-};
+
+
+
