+++ __mocks__/course.ts
@@ -1,7 +1,6 @@
-import { Challenge, Course, LearningModule, Material } from "@/types/course";
+import { Course, LearningModule, Material } from "@/types/course";
 import { mockTrailer } from "./bounty";
-import { reward } from "./reward";
-import { community, metadata } from "./community";
+
 
 export const Introduction = {
   text: "course intro",
@@ -11,6 +10,7 @@ export const mockCertificateData = {
   narrative: "course certificate",
   icon: "certificate icon",
 };
+
 export const Rubric = {
   id: "id",
   ref: "rubric references",
@@ -132,38 +132,4 @@ export const Format = {
   disclaimer: true,
 };
 
-export const challenge: Challenge = {
-  id: "challenge",
-  ref: "challenge ref",
-  created_at: new Date("2022-05-01T12:00:00Z"),
-  updated_at: new Date("2022-05-01T12:00:00Z"),
-  hint: "Hint",
-  name: "challenge name",
-  format: Format,
-  description: "challenge description",
-  course: mockCourse,
-  type: "course",
-  isTeamChallenge: false,
-  objectives: ["objectives 1", "Objectives 2", "Objectives 3", "Objectives 4"],
-  threshold: 8,
-  community: community,
-  reviewTime: 9,
-  metadata: metadata,
-  level: 58,
-  courses: [mockCourse],
-  learningModules: [mockLearningModule],
-  expiresAt: "2025",
-  reward: reward,
-  certificateIcon: "certificate",
-  certificateData: mockCertificateData,
-  ratingCriteria: [mockRatingCriteria],
-  timestamp: 6,
-  maxPoints: 299,
-  minPoints: 9,
-  rewards: [reward],
-  feedbacks: {},
-  feedbackInfo: [{}],
-  bestSubmissions: [{}],
-  teamLimit: 5,
-  isHackathon: false,
-};
+
