+++ __mocks__/reward.ts
@@ -11,8 +11,8 @@ export const reward = {
     amount: 10,
     timestamp: 10,
     distribution: {
-        first: "first distribution",
-        second: "second distribution",
-        third: "third distribution"
+        first: 1,
+        second: 2,
+        third: 3
     }
 }
