+++ __tests__/components/sections/challenges/Header.test.tsx
@@ -2,9 +2,7 @@ import ChallengeHeader from "@/components/sections/challenges/Header";
 import "@testing-library/jest-dom";
 import { screen } from "@testing-library/react";
 import { renderWithRedux } from "../../../../__mocks__/renderWithRedux";
-import { challenge } from "../../../../__mocks__/course";
-import { useSelector } from "react-redux";
-// import { Course } from "../../../../__mocks__/course";
+import { challenge } from "../../../../__mocks__/challenge";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -17,15 +15,8 @@ jest.mock("next/router", () => ({
   }),
 }));
 
-jest.mock("react-redux", () => ({
-  ...jest.requireActual("react-redux"),
-  useSelector: jest.fn(),
-}));
 
 describe("ChallengeHeader", () => {
-  beforeEach(() => {
-    (useSelector as jest.Mock).mockImplementation((callback) => callback({ challenges: { current: challenge } }));
-  });
   it("should render the challenge header with section", () => {
     renderWithRedux(<ChallengeHeader />);
     const challengeheader = screen.getByTestId("challengeHeaderId");
