+++ __tests__/components/sections/challenges/Rubric.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { screen } from "@testing-library/react";
 import { renderWithRedux } from "../../../../__mocks__/renderWithRedux";
 import RubricHeader from "@/components/sections/challenges/Rubric";
-import { RatingCriteria, Rubric } from "../../../../__mocks__/course";
+import { Rubric, mockRatingCriteria } from "../../../../__mocks__/course";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -23,8 +23,8 @@ describe("Rubric", () => {
   });
 
   it("should render the Rubric with ratings", () => {
-    renderWithRedux(<RubricHeader ratingCriteria={[RatingCriteria]} selected={[Rubric]} />);
-    const rubricHeaderName = screen.getByText(RatingCriteria.name);
+    renderWithRedux(<RubricHeader ratingCriteria={[mockRatingCriteria]} selected={[Rubric]} />);
+    const rubricHeaderName = screen.getByText(mockRatingCriteria.name);
     expect(rubricHeaderName).toBeInTheDocument();
     const selectedRubric = screen.getByText(Rubric.text);
     expect(selectedRubric).toBeInTheDocument();
